backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
requests>=2.32.0
